alert-notifier-system/requirements.txt
kafka-python==2.3.0
prometheus-client
